# api/openapi/atm-directory-service.yaml
openapi: 3.0.3
info:
  title: ATM Directory Service API
  version: 0.1.0
  description: Public API to list ATM locations and status.
servers:
  - url: https://api.example.com/open-finance/v1
paths:
  /atms:
    get:
      summary: List ATMs
      security: []
      operationId: listAtms
      tags:
        - ATMs
      parameters:
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/AuthorizationOptional'
        - $ref: '#/components/parameters/Latitude'
        - $ref: '#/components/parameters/Longitude'
        - $ref: '#/components/parameters/Radius'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: ATM list
          headers:
            X-FAPI-Interaction-ID:
              schema:
                type: string
            X-OF-Cache:
              schema:
                type: string
            ETag:
              schema:
                type: string
            Cache-Control:
              description: Always `no-cache`. Any cache may store the response but must revalidate it with If-None-Match before reuse, so a reused response never carries another caller's X-FAPI-Interaction-ID.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtmListResponse'
        '304':
          description: Not modified
          headers:
            ETag:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: >-
            Too many requests. Sent by the API gateway, not by the service. The mesh ingress
            rate-limits this anonymous route with a shared 50 requests per second per gateway
            pod for all callers (token bucket of 100, refilled with 50 every second): it is not
            a per-client quota, so one busy caller can exhaust it for everyone. No retry delay
            is advertised; retry with back-off (exponential, with jitter).
          headers:
            x-fbx-rate-limited:
              description: Set by the gateway when it rejected the request for the rate limit.
              schema:
                type: string
                enum:
                  - 'true'
        '503':
          description: Directory store temporarily unavailable; retry after the given number of seconds
          headers:
            Retry-After:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
    InteractionId:
      name: X-FAPI-Interaction-ID
      in: header
      required: true
      schema:
        type: string
    AuthorizationOptional:
      name: Authorization
      in: header
      required: false
      description: Not needed (public data). When sent, it must use the Bearer or DPoP scheme.
      schema:
        type: string
    Latitude:
      name: lat
      in: query
      required: false
      description: Latitude of the search centre. Must be sent together with `long`.
      schema:
        type: number
        format: double
    Longitude:
      name: long
      in: query
      required: false
      description: Longitude of the search centre. Must be sent together with `lat`.
      schema:
        type: number
        format: double
    Radius:
      name: radius
      in: query
      required: false
      description: Search radius in km, greater than 0 and at most 50; defaults to 5. Ignored without `lat` and `long`.
      schema:
        type: number
        format: double
        exclusiveMinimum: true
        minimum: 0
        maximum: 50
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      schema:
        type: string
  schemas:
    AtmListResponse:
      type: object
      properties:
        Data:
          type: object
          properties:
            ATM:
              type: array
              items:
                $ref: '#/components/schemas/Atm'
        Links:
          type: object
          properties:
            Self:
              type: string
              description: Relative link to this result (path plus the lat, long and radius that were sent), for example `/open-finance/v1/atms?lat=25.2048&long=55.2708&radius=2.0`. Never contains a host name.
        Meta:
          type: object
          properties:
            TotalRecords:
              type: integer
    Atm:
      type: object
      properties:
        AtmId:
          type: string
        Name:
          type: string
        Status:
          type: string
        Latitude:
          type: number
          format: double
        Longitude:
          type: number
          format: double
        Address:
          type: string
        City:
          type: string
        Country:
          type: string
        Accessibility:
          type: string
        Services:
          type: array
          items:
            type: string
        Currency:
          type: string
          description: ISO 4217 code of the currency the ATM dispenses.
          pattern: '^[A-Z]{3}$'
        UpdatedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        interactionId:
          type: string
        timestamp:
          type: string
          format: date-time